refactor: renamed files to follow file name format
