+++ lib/playing_cards.dart
@@ -1,8 +1,8 @@
-export 'src/model/card-value.dart';
+export 'src/model/card_value.dart';
 export 'src/model/suit.dart';
-export 'src/model/playing-card.dart';
-export 'src/views/playing-card-view.dart';
-export 'src/views/playing-card-view-style.dart';
+export 'src/model/playing_card.dart';
+export 'src/views/playing_card_view.dart';
+export 'src/views/playing_card_view_style.dart';
 export 'src/model/deck.dart';
-export 'src/views/flat-card-fan.dart';
-export 'src/util/card-aspect-ratio.dart';
+export 'src/views/flat_card_fan.dart';
+export 'src/util/card_aspect_ratio.dart';
