+++ lib/src/model/card_value.dart
