+++ lib/src/views/flat_card_fan.dart
