+++ lib/src/views/rank_card_center.dart
